refactor: Move validateArchivedQueryParams into list-workflows/helpers
